bitmask pointer

+++ crates/wamrtime_avm/src/lib.rs
@@ -97,7 +97,8 @@ extern "C" fn avm_get_global_uint(
         "avm_get_global_uint called with app: {}, key_ptr: {}, key_len: {}",
         app, key_ptr, key_len
     );
-    let key = wamrtime::runtime::get_wamr_slice(exec_env, key_ptr as u64, key_len as u64);
+
+    let key = wamrtime::runtime::get_wamr_slice(exec_env, key_ptr, key_len);
     let args: GetGlobalUintArgs = [app, key.as_ptr() as u64, key.len() as u64];
     let mut ret: GetGlobalUintRet = [0];
     let num_returns = unsafe {
@@ -129,7 +130,7 @@ extern "C" fn avm_set_global_uint(
         "avm_set_global_uint called with app: {}, key_ptr: {}, key_len: {}, value: {}",
         app, key_ptr, key_len, value
     );
-    let key = wamrtime::runtime::get_wamr_slice(exec_env, key_ptr as u64, key_len as u64);
+    let key = wamrtime::runtime::get_wamr_slice(exec_env, key_ptr, key_len);
     let args: SetGlobalUintArgs = [app, key.as_ptr() as u64, key.len() as u64, value];
     unsafe {
         dispatcher(
